Move under Docker.Client namespace
We re-export everything under Docker so a user can just use
"import Docker" and get everything that they need. But the core
of the library lives under the Docker.Client namespace.
